Done problem one of day 1
The problem is solved by going through each character on each line and taking
the first number and the last number that appears, concatenating them into a two-digit number.
Then, once you are done with all the lines, you add up all the numbers found. And that is the result